Polish up compilation issues
